Move direct access out of PETScWrappers::VectorBase.
This is a breaking change, but a small one - no one should be directly using
this class anyway.

This is part of a sequence of refactorings which will enable us to make a PETSc
BlockVector act like a Vec (by inheriting from VectorBase) and also clean up the
PETSc interfaces (by making VectorBase a VectorSpaceVector).

For the first case, we really cannot use internal::VectorReference with block
vectors since the PETSc block vector implementation of VecGetArray() is a very
expensive operation (it copies every block into a single contiguous array).